update Bill: fix after review
